+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Added
 - [#34](https://github.com/tweag/genealogos/pull/34) implements a web-based GUI for Genealogos
 - [#36](https://github.com/tweag/genealogos/pull/36) include nixtract's new narinfo information
+- [#38](https://github.com/tweag/genealogos/pull/38) display nixtract's status information when running